initial cut at a test app with a sample
