+++ include/orc/orc.hpp
@@ -1,0 +1,17 @@
+// Copyright 2021 Adobe
+// All Rights Reserved.
+//
+// NOTICE: Adobe permits you to use, modify, and distribute this file in accordance with the terms
+// of the Adobe license agreement accompanying it.
+
+#pragma once
+
+// stdc++
+#include <filesystem>
+#include <vector>
+
+/**************************************************************************************************/
+
+int orc_process(const std::vector<std::filesystem::path>&);
+
+/**************************************************************************************************/
