+++ test/battery/vtable/odrv_test.toml
@@ -1,0 +1,28 @@
+sources = [
+    "one.cpp",
+    "two.cpp"
+]
+
+[compile_flags]
+
+common = [
+    # compiler flags for all source files
+]
+
+one = [
+    # compiler flags specifically for one.cpp (the .cpp is implied / omitted)
+]
+
+two = [
+    # compiler flags specifically for two.cpp (the .cpp is implied / omitted)
+]
+
+[orc_test_flags]
+
+# Preserve the object files made by the test app.
+# Default: false
+preserve_object_files = true
+
+[orc_flags]
+
+# Flags specific for ORC
