+++ test/battery/vtable/one.cpp
@@ -1,0 +1,9 @@
+struct object {
+    virtual int optional() const;
+
+    virtual int api() const;
+};
+
+int one(const object& o) {
+    return o.api();
+}
