+++ test/battery/vtable/two.cpp
@@ -1,0 +1,7 @@
+struct object {
+    virtual int api() const;
+};
+
+int two(const object& o) {
+    return o.api();
+}
